keras/server/requirements.txt
numpy
scipy
h5py
tensorflow-gpu
keras
flask
requests
redis
Pillow
